moved the proxyfix out of __init__ to its own location
